Fixed too-fast state machine bug
Fixed a bug where the state machine was ticking by too quickly after
~18 hours of run time. This was caused by saving the value of the
last system tick into a 16 bit word instead of a 32-bit word.

+++ power_fw/src/main.c
@@ -61,7 +61,7 @@ static void check_charge_enable(void)
 int main( void )
 {
 	uint8_t oscval;
-	uint16_t last_tick = 0;
+	uint32_t last_tick = 0;
 	uint32_t system_ticks;
 
 	// Make sure DIV8 is not selected
